maint: supply stub fog config to make tests pass...
We need to supply an AWS secret set for the Fog gem to run, even mocked, in
the test suite.  They don't need to *signify* anything, though, in that they
any just there to satisfy a "mandatory argument" test in the EC2 provider
before it hits the stubs.

So, instead of depending on the user to supply any appropriate configuration
file we provide one: now the tests pass automatically, and everyone is happy.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>