Minor update to README regarding puppet-module
This change adds information to the README file about how to use the
Puppet Module tool to build this module into a package and install the
resulting package into Puppet's modulepath.

Using this method, the end user does not need to modify their RUBYLIB
environment variable and it more closely follows the desired end user
experience.