(#8286) Warn if payload is not a gz file
It should warn b/c otherwise its not clear that the
file must be a tar.gz or a tgz file.

In line 313 of CLoudPack you can see the following text:

'tar -xvzf /tmp/puppet.tar.gz -C /tmp',

this text displays that it actually is a requirement for PE
installations.