(#8348) Add a fingerprint action
This change moves the fingerprint logic out of the create action into a
discrete fingerprint action.

This change is intended to reduce the time required to create multiple
EC2 instances using the create action.

In addition, there is no AWS API to obtain the SSH host key fingerprint,
so the strategy of scraping the AWS console data which is available via
API calls is time consuming and unreliable.

This change also makes an effort to call out that we're making a best
effort and provide steps to help the end user select an AMI that does
actually provide the fingerprint in a securely obtainable fashion.

The tests related to the fingerprinting have been marked as pending
since I currently don't have a good way to mitigate the polling required
of the Mocked responses.

Reviewed-by: Dan Bode