Add retry loop to install action
As part of the code review, Dan gave feedback that moving the
fingerprint logic out of the create action requires additional logic to
handle the race condition of attempting an install action before the
newly launched system is ready.

This small change adds a retry loop to the install action.  This is a
mitigating step until we're able to fully address and clean up the
install action.

Reviewed-by: Dan Bode