(#7508) Fixed broken installer test
Fixed test that verified the installer could find
scripts in the cloudpack source.

The file test.erb is no longer a valid script,
fixed the test to use a valid script.

The test was expecting cloudpack install scripts
to be in the incorrect directory.

Reviewed-by: Jeff McCune

(Deliberate removal of invalid tests are OK)