(#8204) Add pending spec test for s3 script
The S3 install script requires the installer-answers option to be set.
This option is validated inside of the install action itself which has
just been refactored by Dan.

This test is pending Dan's refactor being merged.  It should then be
modified to work in the way Dan has refactored the install action.