(#8369) Mocked more Fog::SSH methods
Mock the upload method of SCP and the run method
of SSH by default to make testing easier.

This allows us to assume the methods are mocked
by default so that we can just consentrate on what
we are testing.

This is probably a patch that needs to be
submitted to Fog.

Reviewed-by: Jeff McCune