(#8402) list action should display status
This change updates the list action to output more than a simple list of
DNS names.  This wasn't ideal for terminated instances because they no
longer have DNS names associated with them.

The output now has the following information in a Hash of Hashes data
structure:

    id: { id, dns_name, created_at, state }

The spec tests have been updated to reflect this change in data
structure from an array to a hash of hashes.

Reviewed-by: Dan Bode