(#8204) Refactor ssh execute method
This change moves all SSH executions into a new method called
ssh_remote_execute.  This method uses Net::SSH directly and configures a
callback to read data from the standard output of the remotely executing
command.

As data is received, we buffer the data until a newline is available
then use Puppet.debug() to display the output to the end user as early
as possible.  This provides a nice compromise between long running
commands with no feedback and spamming the user with chopped up lines as
it streams back using Puppet.debug.

The spec tests have been adjusted to mock the calls wrapping the
Net::SSH and Fog::SSH, Fog::SCP methods.

Reviewed-by: Dan Bode