(#8204) Add ssh connection robustness
This change addresses Dan's feedback that there are a number of
conditions where the install action does not properly retry.

Before the created system has an address on the internet, we expect a
Errno::ETIMEDOUT exception.  Once the kernel is booted but the SSH
service is not started, we expect a Errno::ECONNREFUSED exception.

After that, there's a chance there are authentication failures until the
init script places our SSH public key in the right place.  These should
come back immediately so we sleep for a short amount of time.

Reviewed-by: Dan Bode