(#8513) Make classifcation of node optional
We should not require that users use the Dashboard
for node classification on order to bootstrap
their nodes.

Classification with the Dashboard is only
performed when node_group is set.

Node group has also been updated to not be
required.

Unit tests that expect node_group to be required
have been removed.