(#9442) Rename EC2 specific actions to node_aws subcommand
The behavior before this change is that EC2 and "Puppet Node" actions
are inter-mixed in the "puppet node" subcommand.  With the introduction
of vmware support we've decided (in #9325) to move platform specific
actions to their own sub-commands.

This change moves the bootstrap, create, fingerprint, list and terminate
actions to the puppet node_aws subcommand.

The subcommand node_ec2 is not being used because of bug #9443.  Instead
we're using node_aws.

The rspec tests have been updated and verified.