(#9590) Add ProgressBar class to draw progress bars
This patch adds the Pupept::CloudPack::ProgressBar class to Cloud
Provisioner.  Without this patch, CP doesn't really have a consistent
and good way to display progress bars to the end user.

This is groundwork to improve the UX and UI of the create action based
on feedback from Randall and Luke on the tech mailing list this past
week.