(#9585)(UX) Add list_keynames action, rename keypair to keyname
This patch addresses a usability concern James brought up with me.

Without this patch the behavior of Cloud Provisioner is to describe the
--keypair and --keyfile option in a way that does not clearly identify
the difference between the two.

Even more confusing, the --keypair option expects a very specific and
limited set of user supplied information.

This patch addresses both of these concerns with the following changes
in behavior:  First, the --keypair option is renamed to --keyname and
the summary and description are updated to clearly indicate this is the
name (label) of the key as shown in the AWS console.  This hopefully
disambiguates the option from the --keyfile option which specifically
expects a local filesystem path.

In addition, this patch adds a list_keynames action which lists out the
available keys just as the AWS console would show you.  This hopefully
will help end users quickly identify the value they want to supply when
creating a new AWS machine instance.

The output looks like:

    [root@centos ~]# puppet node_aws list_keynames
    cody (58:c6:4f:3e:b5:51:e0:ec:49:55:4e:98:43:8f:28:f3:9a:14:c8:a3)
    jeff (6e:b6:0a:27:5b:67:cd:8b:47:74:9c:f7:b2:b0:b9:ab:3a:25:d0:28)
    matt (4b:8c:8d:a9:e5:88:6a:47:b7:8b:97:c5:77:e7:b7:6f:fd:b9:64:b3)

The help for create now gives an indication of how to figure out what to
provide the mandatory --keyname option:

    --keyname= - The AWS SSH key name as shown in the AWS
                 console. Please see the related list_keynames
                 action.