(#9608) Add ssh-agent support via --keyfile=agent
The behavior before this patch is to require the end user to use a
private key file on disk.  This is undesirable in the situation where
the end user is logging into a "controller" node to invoke Cloud
Provisioner.  This may be a common situation given a Mac or Windows
laptop and Puppet or Puppet Enterprise installed on a remote system.

This patch allows the user to specify --keyfile agent to satisfy the
requirement of specifying this option.  In this situation, Net::SSH will
default to using the SSH_AUTH_SOCK environment variable to perform
public key authentication.

Test coverage has also been added to validate the exception when
SSH_AUTH_SOCK is not set properly and --keyfile agent is provided.

Finally, some small whitespace only changes are introduced in this
patch.