(#9608) Ensure --keyfile agent is always a downcase value
The behavior before this patch allows the end user to specify
--keyfile=agent in a case insensitive way.  The value wasn't actually
downcased though, which makes checking for "agent" difficult in other
sections of the code since it may be mixed case.

This patch addresses the issue by forcing the value of "agent" to always
be lower case.  Therefore, if the user specifies --keyfile=AGENT the
code can assume it will always be the literal string 'agent'