(#9621) Fix misuse of --certname option in install action
The behavior of the install action without this patch is that it still
assumes it should set options[:certname] to a GUID.  This is a problem
for the user experience and does not really make sense now that we
support installation on platforms other than AWS.

The install action now returns a hash with a 'status' key indicating if
the installation succeeded or not.  This is a bit of a misnomer because
a problem may have arisen in the installation, but this should be
reported as a separate error if the SSH command returns non zero.

This also addresses Dan's comment in the original pull request that we
shouldn't be setting options[:certname] now that we have the
--puppetagent-certname command line option