(#9679) Check if node exists before creating
Previously, any attempt to interact with an existing
node resulted in confusing warning about bad
response codes.

This patch creates a more generic method for
checking that a certain call returns a specific
exit code.

It also adds code so that we check for the existance
of a node before we attempt to create it in the
dashboard.