(#10206) Use exception class as key
The code was incorrectly using the instance of the
exception instead of the class constant. This was
causing it to try to pass nil to the info method
which resulted in an error.

This commit uses the class constant instead of the
instance.

It also removes some trailing whitespace and cleans
up the exception messages.