(#10184) Fix 2 failing specs for CP installer method
Two specs were not updated when the switch was made to use the
puppet-community install script as the default installation script used
by the CP bootstrap and install commands.

The errors from the failing specs where a bit misleading, they
suggested there was a problem with prepending `sudo` to commands
generated by CP bootstrap and install commands. The real problem was due
to the fact that both failing specs where expecting the install command
generated by CP to use `gems.sh` instead of the new default
`puppet-community.sh`.