(#10377) Add --enc-ssl option and https support to classify action
This patch adds the --enc-ssl option to the classify action.  Without
this patch, the classify action only speaks to the Puppet Dashboard
using plain HTTP.  When the end user specifies this command line option
the classify action will use HTTPS rather than HTTP to communicate with
the Dashboard.

It is important to note this patch does not include a post validation
check of the HTTPS connection.  This means this patch provides
encryption, but not assurances of validity of the remote service.

Spec tests are updated in a separate commit.