(#10377) Fix backtrace in classify when connection refused
The behavior without this patch is that the classify action would use
Kernel#exit(1) after catching a connection refused exception.  This
behavior is a problem because we throw away the backtrace and relevant
context regarding the error.

This patch fixes the problem by catching the exception, displaying a
more helpful error message to the user, then throwing a new exception
with the original backtrace and error message.