(#10655) Add positional arguments to puppet help node
Without this patch puppet help node {install,classify} and puppet help
node_aws {fingerprint,terminate} all require positional arguments but
did not mention this in the inline help.

This patch fixes the problem by adding the arguments attribute to each
face action in cloud provisioner that requires a positional argument to
be specified.