Add do_in_progress_bar method to Puppet::CloudPack
Without this patch, the do_in_progress_bar method is specific to a
single cloud provisioner provider.  This patch copies that method into
our main Puppet::CloudPack module, making it available to all of our
current and future cloud provisioner providers.

This patch will not affect/change any currently existing behavior.
The patch will simply lay out the ground work for future patches to
take advantage of the do_in_progress_bar functionality.

One change was made as part of the move.  This patch removes the
fail over condition that would set the progress bar's timeout to Fog's
default timeout and leaves this up to the Puppet face utilizing this
method.  This change is being made so that future provisioners won't
have an arbitrary requirement on the Fog gem.

This patch does NOT provide test coverage for the added method.  There
are currently NO tests written for this method available in its
original repository.