(#11304) Add custom tag support
Previous to this commit, it was not possible for the
user to specify custom EC2 tags for instances.
The only tag generated had a key of 'Created by'
with a value of 'Puppet'.

Users commonly use tags to help identify and organize
instances within their organization. This commit
provides the capability for users to specify the tags
during the create action.

It does so by specifying a --tags option that takes a
value in the form of 'tag1=value,tag2=value'.
Currently the ',' cannot be escaped.  The create_tags
function has been rewritten to accept a third
parameters that is a hash of the tags to generate.

Spec tests to test the tags hash is being generated
correctly are included.