(#12267) Make filter for termination configurable
Previously, the terminate command hard-coded
dns-name as the filter used to retrieve the
server to be terminated.

In the nova implementation of the ec2 api, the only
filter that can be used for the describe_instances
action is instance-id.

This commit creates another option: 'terminate_id'
that is used to specify the field used to filter
servers to be terminated.

The default for this value is set to 'dns-name',
ensuring backwards compatibility. This default
is also set in the terminate method (as opposed to
by merge_default_options) b/c it is a termination
specific option.