(#12228) Add endpoint option to creation of fog compute
This commit adds an option called :endpoint to
the creation of the fog compute object.

Endpoint refers to the url of the api server that
implements the ec2 api. In the case of aws, the
endpoint is derived based on the selected region (
b/c each region in aws has a hard-coded api url).
In the case of nova compute, the endpoint is the
internal url of your ec2 api server.

I verified that Fog correctly ignores any
endpoint set to nil in 1.1 (and master)