(#7220) Refactor to use default_to for option defaults
Option defaults are currently handled by the
method merge_default_options. This method is called
throughout the code. It hard-codes three options to
be used as default when no explicit option is provided.
  - region, platform, install_script

This approach is problematic b/c it will set
options even if the option was not explicitly specified
as on option for the face/action.

This is problematic for openstack b/c it will
still specify the region option in the fog connection
even though region is not currently a supported option.

This refactor removes the merge_default_options
method from Puppet::CloudPack. It also removes all
invocations of the method.

Each of the option default is set using the
default_to method which is passed to the option's
block.

I verified that default_to is called for every option
before before_action is called for any option. This
seemed to be one of the main concerns that lead to
the default_to method not being used in the first
place (based on inline comments)

This commit also ensure that add_platform_option is
called before add_region_option so that the
platform option can correctly catch errors.