(#12352) Reset Fog::Compute::AWS::Mock objects
Mock servers persist on the class instance of
Fog::Compute::AWS::Mock which means that they
persist beyond the scope in which they are created.

This introduces a coupling between tests that makes
the tests harder to understand and debug.

This commit ensures that all of the data associated
with Fog::Compute::AWS::Mock is reset after each
test. This required cleaning up tests to ensure
that each test created the mock objects that it
needs.