(12717) Don't append to answers file if puppet agent certname already present
Previously, if the `--puppetagent-certname` command line option was
not specified on the command line, then the `puppet-enterprise.erb`
template would always append a `q_puppetagent_certname` entry to the
answers file, regardless if one was already present. This was due to a
typo -- we were checking puppet.answer, but appending to
puppet.answers.