Provide puppet class to manage installation of CP
Previous to this commit, the installation of Cloud Provisioner required
a pluginsync to the agent as well as adding the $vardir to the $RUBYLIB
path. This was due to a long standing bug with Faces that requires any
Face that creates a new Puppet::Application child class to be manually
in the $RUBYLIB path.

This commit resolves this issue but providing a cloud_provisioner class
that installs the node_aws.rb file in the puppet/application directory.
The path to the application directory is determined by a Facter fact
this module also provides, $puppet_install_dir.  The class contains two
parameters:

* **$ensure** - Can be **present**, **absent**, **installed**, and
  **uninstalled**. Defaults to **present**
* $puppet_install_directory - Defaults to value of
  **$puppet_install_dir** fact

**NOTE**: This class is not necessary in Puppet 3.0+ as the Face
Puppet::Application bug has been fixed