Manage Fog credentials
Previous to this commit, in order to set up Cloud Provisioner, the user
was forced to manually manage their ~/.fog credential file.  This commit
provides a cloud_provisioner::credential defined type that manages the
.fog file.  The location of the .fog file is determined by the
**$cloud_provisioner::provisioner_account_home_directory** parameter.

Further, the Class['cloud_provisioner'] has seven new parameters to aid
in the management of the Fog credentials.
* **provisioner_account_home_directory** - The home directory path to
  the CP control user.  Defaults to /root
* **default_aws_access_key_id**
* **default_aws_secret_access_key**
* **default_vsphere_server**
* **default_vsphere_username**
* **default_vsphere_password**
* **default_vsphere_expected_pubkey_hash**