(#15450) Add terminate-id option to node_aws terminate
This commit adds the terminate-id option to the node_aws
terminate command.

This option allows users to select either dns-name (default)
or isntance-id as the identifier to use for termination.

This option had previously been added to Puppet::Cloudpack
terminate method for use with the openstack cloud provisioner
plugin, but not exposed as an option to the face.