retry ssh when Errno::ENETUNREACH is encountered
This commit adds Errno::ENETUNREACH to the list of exceptions
that should result in a retry when encountered during ssh connections.

This exception can indicate that routing to an ip address has not
properly stabilized. I ran into this error when using source NAT
VPN with a cloudstack environment and I verified that the connection
was established if I retried a few times when this error was encountered.