Add floating IP support
This patch is primarily a base for floating IP support using
node_openstack.  This patch allows a user to define --floating-ip
when creating a new instance.  During instance creation the IP pool
will be checked for free IPs before spawning a new instance.  If
there are free IPs the instance will be created then assigned a
floating-ip.  If there are no free IPs an error message will be
displayed informing the user there are currently no free IPs in the
IP pool.

This patch should work with EC2 as well, but since I was not able to
test it I left the option disabled for node_aws.