PE-239: Cloud Provisioner has unused, obsolete `--pe-version` option
The Cloud Provisioner tools added an option to the command, `--pe-version`,
that claimed it determined what version of PE was installed.  In practice,
though, it was unused - nothing in the rest of the codebase consulted it.

This eliminates the option which, while very attractive, is now implemented by
picking the correct installation tarball or script source.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>