PE-247: SSL for ENC communication is now on, and mandatory
Earlier versions of Cloud Provisioner had an option to enable SSL for
communication with the ENC.  This was, by default, off - and included a
recommendation that any time in the last year it should be manually enabled by
the user.

This changes that: not only defaulting to turning on SSL, but making it
mandatory for all ENC communication.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>