PE-242: remove obsolete local AWS type validation
The original AWS provisioning code was careful to validate the set of approved
EC2 instance types locally.  This was great, except that it wasn't kept up
to date.

A better strategy is to simply delegate this to the server: while it takes
longer to give the user an error message, EC2 is clear and effective in
communicating what went wrong, and we can't get much more authoritative than
the upstream service when it comes to actual validation.

Since we don't own changes to this data, and we can't upgrade the valid set in
the field, this seems like the better approach when trading off speed of error
reporting against ability to respond to change in third party systems.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>