Config both 'main' and 'dependencies' apt repos
Previously the apt installation setup only configured the
'main' repository. With Puppet 3.2's hard dependency on
the ruby-rgen module, this causes installation to fail due
to unmet dependencies.

This patch updates the script to configure both the 'main' and
'dependencies' repos so that `puppet node install` will work
on debian-based systems installing Puppet 3.2