PE-834, PE-424: fixes for cloud provisioner PE console interaction
Update the cloud provisioner to use the
Puppet::Network::HTTP::Connection class correctly - a fix for PE-834.
While at it add the --insecure option to the command to make it possible
to connect to PE console hosts without checking their SSL certificates.
Note that that is impossible while using the mentioned
Puppet::Network::HTTP::Connection class for PE console interaction
as it always tries to verify the server certificate.  So when the
--insecure options is specified the Net::HTTP class is used instead
for the interaction.