PE-834, PE-424: fixes for Cloud Provisioner PE console interaction
Fix for PE-834:
Update the CP to work correctly with the new object type:
  Puppet::Network::HTTP::Connection
returned by the:
  Puppet::Network::HttpPool.http_instance
method. Note that previously the method used to return objects of type:
  Net:HTTP

Fix for PE-424:
Since the Puppet::Network::HTTP::Connection performs peer verification
by default, the PE-424 is fixed just by using connection objects of that
type. But as there may be situations when it is not possible to verify
the peer, a new option - "--insecure" - has been added which can be used
to suppress the peer verification.