PE-830: update doc of puppet node install --installer-payload option
This patch updates the inline documentation of the --installer-payload
option of the
  puppet node install
command.
Previously the doc stated that the option's argument must point to
the universal PE install tarball. It turns out this is not necessary:
A platform specific tarball can be specified as long as it supports
the platform of the node it is supposed to be installed on.
The doc has been updated to reflect that fact.