PE-838: default to installing PE during bootstrap
This patch changes the default value of the --install-script
option of the:
  puppet node_aws bootstrap
command to 'puppet-enterprise'.
While at it the option verification is improved so that missing
options are detected early rather than halfway through the
bootstrap process.
Further the vocabulary of options' help texts and related error
messages is unified.