PE-1178: --security-group option handling improvements
This patch improves the handling of the --security-group option
of the
  puppet node_aws create
command.
Previously it was only possible to specify the security groups by
their names, newly it is possible to specify them by both IDs and
names interchangeably. Internally they are always converted to IDs
and as such passed to the Puppet::CloudPack.create_server method
in the :security_group_ids option. (Previously the names of the
groups were passed in the :security_group option to that method.)