Add initial GCE support: Google API authentication
This implements the first phase of Google Compute Engine support to Cloud
Provisioner, by implementing authentication support.  This uses the OAuth2
installed application workflow, which seems to be the best match to what is
offered by the cloud console, and our local needs.

This implements token storage, refresh, and a web browser based access
request process.  This is initially entered by the user through the `register`
action on the `node_gce` face.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>