Support `google_compute_engine` ssh key by default
The upstream gcutils tools default to the `google_compute_engine` ssh key;
this adapts our code to look for that first, and default to it if present.

If it does not exist we fall back to the normal `id_rsa` SSH credentials,
which follows the common practice of the rest of Cloud Provisioner.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>