Make `zone` argument optional for listing GCE instances
The GCE list operation can provide a per-zone list, or an aggregate list.
The later shows all nodes -- and was broken by earlier changes that unified
adding options to things, by setting a default value for zone on the
list action.

This fixes that, by extending the API for adding options to support an
optional default value for zone.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>