PE-1620: handle class names in retry_exception hash correctly
Previously the +Puppet::CloudPack::Utils.retry_action+ method
was able to handle retry_excpetion hash in which the exceptions
were specified by strings (or symbols). But it failed when the
--debug option was specified at the same time because it tried
to lookup the caught exception description in the retry_exception
hash using the exception class object as a key to the hash instead
of the class name string (or symbol).
This was fixed so now the method can handle exceptions specified
as string (or symbols) correctly under all circumstances.