(maint) Pin to google-api-client version
Prior to this commit, we had been pulling in the latest version of
google-api-client. This is generally fine, except that gem has a runtime
dependency `mime-types >= 1.6`, which has a runtime dependency
`mime-types-data ~> 3.2015` which cannot be installed because it
requires Ruby >= 2.0. We need to be able to run these tests on Ruby
1.9.3, so we need to pin to an older version of google-api-client that
doesn't have a dependency chain that requires Ruby 2.0.